+++ lib/pkgr/cli.rb
@@ -31,7 +31,7 @@ module Pkgr
 
     method_option :buildpack,
       :type => :string,
-      :desc => "Custom buildpack to use"
+      :desc => "Custom buildpack to use. May be a single string or a list."
     method_option :buildpack_list,
       :type => :string,
       :desc => "Specify a file containing a list of buildpacks to use (--buildpack takes precedence if given)"