+++ lib/pkgr/distributions/base.rb
@@ -76,12 +76,10 @@ module Pkgr
 
       # Returns a list of Buildpack objects
       def buildpacks
-        custom_buildpack_uri = config.buildpack
-        if custom_buildpack_uri
-          uuid = Digest::SHA1.hexdigest(custom_buildpack_uri)
-          [Buildpack.new(custom_buildpack_uri, :custom, config.env)]
+        if config.buildpacks
+          [:custom, config.buildpacks.map { |url| Buildpack.new(url, :custom, config.env) }]
         else
-          load_buildpack_list
+          [:builtin, load_buildpack_list]
         end
       end # def buildpacks