+++ lib/pkgr/env.rb
@@ -2,6 +2,10 @@ module Pkgr
   class Env
     attr_reader :variables
 
+    def self.from_export(file)
+      new(File.read(file).split("\n").map { |line| line.sub(/^export\s+/, "") })
+    end
+
     def initialize(variables = nil)
       @variables = variables || []
     end